Reorganize LedgerDB
