+++ ouroboros-consensus/ouroboros-consensus.cabal
@@ -162,7 +162,6 @@ library
     Ouroboros.Consensus.Ledger.Tables.Basics
     Ouroboros.Consensus.Ledger.Tables.Combinators
     Ouroboros.Consensus.Ledger.Tables.Diff
-    Ouroboros.Consensus.Ledger.Tables.DiffSeq
     Ouroboros.Consensus.Ledger.Tables.MapKind
     Ouroboros.Consensus.Ledger.Tables.Utils
     Ouroboros.Consensus.Mempool
@@ -228,18 +227,16 @@ library
     Ouroboros.Consensus.Storage.ImmutableDB.Impl.Iterator
     Ouroboros.Consensus.Storage.ImmutableDB.Impl.Parser
     Ouroboros.Consensus.Storage.ImmutableDB.Impl.State
-    Ouroboros.Consensus.Storage.ImmutableDB.Impl.Stream
     Ouroboros.Consensus.Storage.ImmutableDB.Impl.Types
     Ouroboros.Consensus.Storage.ImmutableDB.Impl.Util
     Ouroboros.Consensus.Storage.ImmutableDB.Impl.Validation
     Ouroboros.Consensus.Storage.ImmutableDB.Stream
     Ouroboros.Consensus.Storage.LedgerDB
     Ouroboros.Consensus.Storage.LedgerDB.API
-    Ouroboros.Consensus.Storage.LedgerDB.Impl.Args
-    Ouroboros.Consensus.Storage.LedgerDB.Impl.Common
-    Ouroboros.Consensus.Storage.LedgerDB.Impl.Init
-    Ouroboros.Consensus.Storage.LedgerDB.Impl.Snapshots
-    Ouroboros.Consensus.Storage.LedgerDB.Impl.Validate
+    Ouroboros.Consensus.Storage.LedgerDB.Forker
+    Ouroboros.Consensus.Storage.LedgerDB.Args
+    Ouroboros.Consensus.Storage.LedgerDB.TraceEvent
+    Ouroboros.Consensus.Storage.LedgerDB.Snapshots
     Ouroboros.Consensus.Storage.LedgerDB.V1.Args
     Ouroboros.Consensus.Storage.LedgerDB.V1.BackingStore
     Ouroboros.Consensus.Storage.LedgerDB.V1.BackingStore.API
@@ -247,17 +244,16 @@ library
     Ouroboros.Consensus.Storage.LedgerDB.V1.BackingStore.Impl.LMDB
     Ouroboros.Consensus.Storage.LedgerDB.V1.BackingStore.Impl.LMDB.Bridge
     Ouroboros.Consensus.Storage.LedgerDB.V1.BackingStore.Impl.LMDB.Status
-    Ouroboros.Consensus.Storage.LedgerDB.V1.Common
     Ouroboros.Consensus.Storage.LedgerDB.V1.DbChangelog
-    Ouroboros.Consensus.Storage.LedgerDB.V1.Flush
     Ouroboros.Consensus.Storage.LedgerDB.V1.Forker
-    Ouroboros.Consensus.Storage.LedgerDB.V1.Init
+    Ouroboros.Consensus.Storage.LedgerDB.V1.DiffSeq
+    Ouroboros.Consensus.Storage.LedgerDB.V1
     Ouroboros.Consensus.Storage.LedgerDB.V1.Lock
     Ouroboros.Consensus.Storage.LedgerDB.V1.Snapshots
     Ouroboros.Consensus.Storage.LedgerDB.V2.Args
-    Ouroboros.Consensus.Storage.LedgerDB.V2.Common
+    Ouroboros.Consensus.Storage.LedgerDB.V2.Forker
     Ouroboros.Consensus.Storage.LedgerDB.V2.InMemory
-    Ouroboros.Consensus.Storage.LedgerDB.V2.Init
+    Ouroboros.Consensus.Storage.LedgerDB.V2
     Ouroboros.Consensus.Storage.LedgerDB.V2.LedgerSeq
     Ouroboros.Consensus.Storage.Serialisation
     Ouroboros.Consensus.Storage.VolatileDB
