+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/Storage/LedgerDB/V1/Forker.hs
@@ -1,94 +1,78 @@
+{-# LANGUAGE DeriveAnyClass #-}
+{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TupleSections #-}
+{-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeFamilies #-}
-{-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE UndecidableInstances #-}
 
 module Ouroboros.Consensus.Storage.LedgerDB.V1.Forker (
-    closeAllForkers
-  , newForkerAtTarget
-  , newForkerByRollback
+    ForkerEnv (..)
+  , closeForkerEnv
+  , implForkerCommit
+  , implForkerGetLedgerState
+  , implForkerPush
+  , implForkerRangeReadTables
+  , implForkerReadStatistics
+  , implForkerReadTables
   ) where
 
-import           Control.ResourceRegistry
 import           Control.Tracer
-import           Data.Functor.Contravariant ((>$<))
 import qualified Data.Map.Strict as Map
 import           Data.Semigroup
 import qualified Data.Set as Set
-import           Data.Word
+import           GHC.Generics (Generic)
 import           NoThunks.Class
 import           Ouroboros.Consensus.Block
+import           Ouroboros.Consensus.Config
 import           Ouroboros.Consensus.Ledger.Abstract
 import           Ouroboros.Consensus.Ledger.SupportsProtocol
 import qualified Ouroboros.Consensus.Ledger.Tables.Diff as Diff
-import           Ouroboros.Consensus.Ledger.Tables.DiffSeq (numDeletes,
-                     numInserts)
-import qualified Ouroboros.Consensus.Ledger.Tables.DiffSeq as DS
-import           Ouroboros.Consensus.Storage.LedgerDB.API as API
-import           Ouroboros.Consensus.Storage.LedgerDB.Impl.Common
+import           Ouroboros.Consensus.Storage.LedgerDB.Forker as Forker
 import           Ouroboros.Consensus.Storage.LedgerDB.V1.Args
 import           Ouroboros.Consensus.Storage.LedgerDB.V1.BackingStore
 import qualified Ouroboros.Consensus.Storage.LedgerDB.V1.BackingStore.API as BackingStore
-import           Ouroboros.Consensus.Storage.LedgerDB.V1.Common
 import           Ouroboros.Consensus.Storage.LedgerDB.V1.DbChangelog
-import           Ouroboros.Consensus.Storage.LedgerDB.V1.Lock
-import           Ouroboros.Consensus.Util
+import           Ouroboros.Consensus.Storage.LedgerDB.V1.DiffSeq (numDeletes,
+                     numInserts)
+import qualified Ouroboros.Consensus.Storage.LedgerDB.V1.DiffSeq as DS
 import           Ouroboros.Consensus.Util.IOLike
-import           Ouroboros.Network.Protocol.LocalStateQuery.Type
+
+{-------------------------------------------------------------------------------
+  Forkers
+-------------------------------------------------------------------------------}
+
+data ForkerEnv m l blk = ForkerEnv {
+    -- | Local, consistent view of backing store
+    foeBackingStoreValueHandle :: !(LedgerBackingStoreValueHandle m l)
+    -- | In memory db changelog, 'foeBackingStoreValueHandle' must refer to
+    -- the anchor of this changelog.
+  , foeChangelog               :: !(StrictTVar m (DbChangelog l))
+    -- | The same 'StrictTVar' as 'ldbChangelog'
+    --
+    -- The anchor of this and 'foeChangelog' might get out of sync if diffs are
+    -- flushed, but 'forkerCommit' will take care of this.
+  , foeSwitchVar               :: !(StrictTVar m (DbChangelog l))
+    -- | Config
+  , foeSecurityParam           :: !SecurityParam
+     -- | Config
+  , foeQueryBatchSize          :: !QueryBatchSize
+  , foeTracer                  :: !(Tracer m TraceForkerEvent)
+  }
+  deriving Generic
+
+deriving instance ( IOLike m
+                  , LedgerSupportsProtocol blk
+                  , NoThunks (l EmptyMK)
+                  , NoThunks (TxIn l)
+                  , NoThunks (TxOut l)
+                  ) => NoThunks (ForkerEnv m l blk)
 
 {-------------------------------------------------------------------------------
   Close
 -------------------------------------------------------------------------------}
 
--- | Will call 'error' if the point is not on the LedgerDB
-newForkerAtTarget ::
-     ( HeaderHash l ~ HeaderHash blk
-     , IOLike m
-     , IsLedger l
-     , StandardHash l
-     , HasLedgerTables l
-     , LedgerSupportsProtocol blk
-     )
-  => LedgerDBHandle m l blk
-  -> ResourceRegistry m
-  -> Target (Point blk)
-  ->  m (Either GetForkerError (Forker m l blk))
-newForkerAtTarget h rr pt = getEnv h $ \ldbEnv ->
-    withReadLock (ldbLock ldbEnv) (acquireAtTarget ldbEnv rr (Right pt)) >>= traverse (newForker h ldbEnv)
-
-newForkerByRollback ::
-     ( HeaderHash l ~ HeaderHash blk
-     , IOLike m
-     , IsLedger l
-     , StandardHash l
-     , HasLedgerTables l
-     , LedgerSupportsProtocol blk
-     )
-  => LedgerDBHandle m l blk
-  -> ResourceRegistry m
-     -- | How many blocks to rollback from the tip
-  -> Word64
-  -> m (Either GetForkerError (Forker m l blk))
-newForkerByRollback h rr n = getEnv h $ \ldbEnv -> do
-    withReadLock (ldbLock ldbEnv) (acquireAtTarget ldbEnv rr (Left n)) >>= traverse (newForker h ldbEnv)
-
--- | Close all open block and header 'Forker's.
-closeAllForkers ::
-     IOLike m
-  => LedgerDBEnv m l blk
-  -> m ()
-closeAllForkers ldbEnv =
-  do
-    forkerEnvs <- atomically $ do
-      forkerEnvs <- Map.elems <$> readTVar forkersVar
-      writeTVar forkersVar Map.empty
-      return forkerEnvs
-    mapM_ closeForkerEnv forkerEnvs
-  where
-    forkersVar = ldbForkers ldbEnv
 
 closeForkerEnv :: ForkerEnv m l blk -> m ()
 closeForkerEnv ForkerEnv { foeBackingStoreValueHandle } = bsvhClose foeBackingStoreValueHandle
@@ -96,136 +80,6 @@ closeForkerEnv ForkerEnv { foeBackingStoreValueHandle } = bsvhClose foeBackingSt
 {-------------------------------------------------------------------------------
   Acquiring consistent views
 -------------------------------------------------------------------------------}
-
-type Resources m l =
-    (LedgerBackingStoreValueHandle m l, DbChangelog l)
-
--- | Acquire both a value handle and a db changelog at the tip. Holds a read lock
--- while doing so.
-acquireAtTarget ::
-     forall m l blk. (
-       HeaderHash l ~ HeaderHash blk
-     , IOLike m
-     , IsLedger l
-     , StandardHash l
-     , HasLedgerTables l
-     , LedgerSupportsProtocol blk
-     )
-  => LedgerDBEnv m l blk
-  -> ResourceRegistry m
-  -> Either Word64 (Target (Point blk))
-  -> ReadLocked m (Either GetForkerError (Resources m l))
-acquireAtTarget ldbEnv rr (Right VolatileTip) =
-    readLocked $ do
-      dblog <- readTVarIO (ldbChangelog ldbEnv)
-      Right . (,dblog) <$> acquire ldbEnv rr dblog
-acquireAtTarget ldbEnv rr (Right ImmutableTip) =
-    readLocked $ do
-      dblog <- readTVarIO (ldbChangelog ldbEnv)
-      Right . (, rollbackToAnchor dblog)
-        <$> acquire ldbEnv rr dblog
-acquireAtTarget ldbEnv rr (Right (SpecificPoint pt)) =
-    readLocked $ do
-      dblog <- readTVarIO (ldbChangelog ldbEnv)
-      let immTip = getTip $ anchor dblog
-      case rollback pt dblog of
-        Nothing     | pointSlot pt < pointSlot immTip -> pure $ Left $ PointTooOld Nothing
-                    | otherwise   -> pure $ Left PointNotOnChain
-        Just dblog' -> Right . (,dblog') <$> acquire ldbEnv rr dblog'
-acquireAtTarget ldbEnv rr (Left n) = readLocked $ do
-      dblog <- readTVarIO (ldbChangelog ldbEnv)
-      case rollbackN n dblog of
-        Nothing ->
-          return $ Left $ PointTooOld $ Just $ ExceededRollback {
-              API.rollbackMaximum   = maxRollback dblog
-            , API.rollbackRequested = n
-            }
-        Just dblog' ->
-           Right . (,dblog') <$> acquire ldbEnv rr dblog'
-
-acquire ::
-     (IOLike m, GetTip l)
-  => LedgerDBEnv m l blk
-  -> ResourceRegistry m
-  -> DbChangelog l
-  -> m (LedgerBackingStoreValueHandle m l)
-acquire ldbEnv rr dblog =  do
-  -- bsvhClose is idempotent, so we let the resource call it even if the value
-  -- handle might have been closed somewhere else
-  (_, vh) <- allocate rr (\_ -> bsValueHandle $ ldbBackingStore ldbEnv) bsvhClose
-  let dblogSlot = getTipSlot (changelogLastFlushedState dblog)
-  if bsvhAtSlot vh == dblogSlot
-    then pure vh
-    else bsvhClose vh >>
-         error (  "Critical error: Value handles are created at "
-                <> show (bsvhAtSlot vh)
-                <> " while the db changelog is at "
-                <> show dblogSlot
-                <> ". There is either a race condition or a logic bug"
-                )
-
-{-------------------------------------------------------------------------------
-  Make forkers from consistent views
--------------------------------------------------------------------------------}
-
-newForker ::
-     ( IOLike m
-     , HasLedgerTables l
-     , LedgerSupportsProtocol blk
-     , NoThunks (l EmptyMK)
-     , GetTip l
-     )
-  => LedgerDBHandle m l blk
-  -> LedgerDBEnv m l blk
-  -> Resources m l
-  -> m (Forker m l blk)
-newForker h ldbEnv (vh, dblog) = do
-  dblogVar     <- newTVarIO dblog
-  forkerKey    <- atomically $ stateTVar (ldbNextForkerKey ldbEnv) $ \r -> (r, r + 1)
-  let forkerEnv = ForkerEnv {
-      foeBackingStoreValueHandle = vh
-    , foeChangelog               = dblogVar
-    , foeSwitchVar               = ldbChangelog ldbEnv
-    , foeSecurityParam           = ledgerDbCfgSecParam $ ldbCfg ldbEnv
-    , foeQueryBatchSize          = ldbQueryBatchSize ldbEnv
-    , foeTracer                  = LedgerDBForkerEvent . TraceForkerEventWithKey forkerKey >$< ldbTracer ldbEnv
-    }
-  atomically $ modifyTVar (ldbForkers ldbEnv) $ Map.insert forkerKey forkerEnv
-  traceWith (foeTracer forkerEnv) ForkerOpen
-  pure $ mkForker h forkerKey
-
-mkForker ::
-     ( IOLike m
-     , HasHeader blk
-     , HasLedgerTables l
-     , GetTip l
-     )
-  => LedgerDBHandle m l blk
-  -> ForkerKey
-  -> Forker m l blk
-mkForker h forkerKey = Forker {
-      forkerClose                  = implForkerClose h forkerKey
-    , forkerReadTables             = getForkerEnv1   h forkerKey implForkerReadTables
-    , forkerRangeReadTables        = getForkerEnv1   h forkerKey implForkerRangeReadTables
-    , forkerGetLedgerState         = getForkerEnvSTM h forkerKey implForkerGetLedgerState
-    , forkerReadStatistics         = getForkerEnv    h forkerKey implForkerReadStatistics
-    , forkerPush                   = getForkerEnv1   h forkerKey implForkerPush
-    , forkerCommit                 = getForkerEnvSTM h forkerKey implForkerCommit
-    }
-
-implForkerClose ::
-     IOLike m
-  => LedgerDBHandle m l blk
-  -> ForkerKey
-  -> m ()
-implForkerClose (LDBHandle varState) forkerKey = do
-    envMay <- atomically $ readTVar varState >>= \case
-      LedgerDBClosed       -> pure Nothing
-      LedgerDBOpen ldbEnv -> do
-        stateTVar
-            (ldbForkers ldbEnv)
-            (Map.updateLookupWithKey (\_ _ -> Nothing) forkerKey)
-    whenJust envMay closeForkerEnv
 
 implForkerReadTables ::
      (MonadSTM m, HasLedgerTables l, GetTip l)
@@ -289,7 +143,7 @@ implForkerRangeReadTables env rq0 = do
          (Ord k, Eq v)
       => SeqDiffMK k v
       -> DiffMK k v
-    prj (SeqDiffMK sq) = DiffMK (Diff.fromAntiDiff $ DS.cumulativeDiff sq)
+    prj (SeqDiffMK sq) = DiffMK (DS.fromAntiDiff $ DS.cumulativeDiff sq)
 
     -- Remove all diff elements that are <= to the greatest given key
     doDropLTE ::
@@ -371,7 +225,7 @@ implForkerGetLedgerState env = current <$> readTVar (foeChangelog env)
 implForkerReadStatistics ::
      (MonadSTM m, HasLedgerTables l, GetTip l)
   => ForkerEnv m l blk
-  -> m (Maybe API.Statistics)
+  -> m (Maybe Forker.Statistics)
 implForkerReadStatistics env = do
     traceWith (foeTracer env) ForkerReadStatistics
     dblog <- readTVarIO (foeChangelog env)
@@ -394,7 +248,7 @@ implForkerReadStatistics env = do
         nDeletes = ltcollapse
                  $ ltmap (K2 . getSum . numDeletes . getSeqDiffMK)
                    diffs
-      pure . Just $ API.Statistics {
+      pure . Just $ Forker.Statistics {
           ledgerTableSize = n + nInserts - nDeletes
         }
   where
